pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["znammikhail <znamenskymihail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.1"
Flask = "^2.2.3"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"